requirements.txt
Package                  Version
------------------------ -------
asgiref                  3.8.1
Django                   5.0.6
django-cors-headers      4.3.1
django-phonenumber-field 7.3.0
django-restframework     0.0.1
djangorestframework      3.15.1
phonenumbers             8.13.38
pillow                   10.3.0
pip                      23.0.1
setuptools               66.1.1
sqlparse                 0.5.0
gunicorn